PotterShoppingCart價格計算測試_第一集買一本_無滿足任何優惠_價格為100(綠燈) PotterShoppingCart價格計算測試_第一集買一本_第二集買一本_滿足95折優惠_價格為190(紅燈)

+++ ShoppingCartLibrary/Calculator.cs
@@ -1,13 +1,16 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading;
+using System.Reflection;
+using ShoppingCartLibrary.Models;
 
 namespace ShoppingCartLibrary
 {
     public static class Calculator
     {
-        public static int GetSum<T>(
-            this IEnumerable<T> shoppingList,Func<T, int> selector)
+        public static double GetSum<T>(
+            this IEnumerable<T> shoppingList,Func<T, double> selector)
         {
             return shoppingList.Sum(selector);
         }
